backend/requirements.txt
asgiref==3.11.1
Django==4.2.28
django-grappelli
django-redis
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
gunicorn==25.0.3
cryptography==46.0.5
lxml==6.0.2
openpyxl==3.1.5
packaging==26.0
psycopg2-binary==2.9.11
PyJWT==2.11.0
python-docx==1.2.0
Pillow==11.1.0
python-dotenv==1.2.1
pyserial==3.5
requests==2.32.3
reportlab==4.2.5
twilio==9.8.1
chardet==5.2.0
whitenoise==6.5.0
sqlparse==0.5.5
typing_extensions==4.15.0
